fix: 関数を全てconに移動
